backend/requirements.txt
# backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
supabase==2.3.0
asyncpg==0.29.0
sqlalchemy==2.0.23

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# OpenAI integration
openai==1.6.0

# Caching and rate limiting
redis==5.0.1
aioredis==2.0.1

# Data processing
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2

# API utilities
httpx==0.25.2
aiofiles==23.2.1

# Monitoring and logging
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Additional utilities
pytz==2023.3
email-validator==2.1.0
python-multipart==0.0.6